requirements.txt
fastapi>=0.110,<1.0
uvicorn>=0.30,<0.40
python-multipart>=0.0.9,<0.1
torch>=2.2,<2.5          # CPU-only is fine unless you need CUDA
torchvision>=0.17,<0.20
numpy>=1.26,<2.3
pillow>=10,<11
pandas==2.1.4
scikit-learn==1.3.2
seaborn>=0.13,<0.14
matplotlib==3.8.2
requests==2.31.0
joblib==1.3.2
mlflow==2.10.0
prometheus-client==0.19.0
pytest==7.4.3
setuptools>=69,<76
